+++ packages/components/components/elvis-accordion/src/react/declarations.d.ts
@@ -1,3 +1,4 @@
 declare module '*.css';
 declare module '*.scss';
 declare module '@elvia/elvis-typography';
+declare module '@elvia/elvis-colors';