+++ packages/components/components/elvis-chip/src/react/declarations.d.ts
@@ -1,4 +1,3 @@
 declare module '*.css';
 declare module '*.scss';
 declare module '@elvia/elvis-colors';
-declare module '@elvia/elvis-typography';